requirements_web.txt
Flask==2.2.3
Flask-SocketIO==5.3.2
python-socketio==5.7.2
python-engineio==4.3.4
eventlet==0.33.3
telethon==1.39.0
requests==2.28.1
python-dotenv==0.20.0 